+++ molecool/__init__.py
@@ -5,7 +5,12 @@ introduction to python packagees
 
 # Add imports here
 from .functions import *
+from .data import atom_colors, atomic_weights
+from .measure import calculate_angle, calculate_distance
+from .molecule import build_bond_list
+from .visualize import draw_molecule, bond_histogram
 
+#from .io import pdb, xyz
 # Handle versioneer
 from ._version import get_versions
 versions = get_versions()
